fixed EXC_BAD_INSTRUCTION (code=EXC_I386_INVOP, subcode=0x0)

+++ Tofu/Password.swift
@@ -12,10 +12,10 @@ final class Password {
     let counter = timeBased ?
       Int64(date.timeIntervalSince1970) / Int64(period) : Int64(self.counter)
     var input = counter.bigEndian
-    let digest = UnsafeMutablePointer<UInt8>.allocate(capacity: algorithm.digestLength)
+    let digest = UnsafeMutablePointer<UInt32>.allocate(capacity: algorithm.digestLength)
     defer { digest.deinitialize() }
     CCHmac(algorithm.hmacAlgorithm, (secret as NSData).bytes, secret.count, &input, MemoryLayout.size(ofValue: input), digest)
-    let bytes = UnsafePointer<UInt8>(digest)
+    let bytes = UnsafePointer<UInt32>(digest)
     let offset = bytes[algorithm.digestLength - 1] & 0x0f
     let number = UInt32(bigEndian: UnsafeRawPointer(bytes + Int(offset)).load(as: UInt32.self)) & 0x7fffffff
     return String(format: "%0\(digits)d", number % UInt32(pow(10, Float(digits))))
